docuvoz/scraper/requirements.txt
requests
beautifulsoup4
python-dotenv
supabase
PyPDF2
python-docx